requirements.txt
annotated-types==0.5.0
anyio==3.7.1
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.6
cssselect==1.2.0
dnspython==2.4.2
email-validator==2.0.0.post2
fastapi==0.101.1
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.12.2
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.3.1
lxml==4.9.3
MarkupSafe==2.1.3
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.25.2
orjson==3.9.4
Pillow==10.0.0
psycopg2==2.9.6
py4j==0.10.9.7
pydantic==2.1.1
pydantic-extra-types==2.0.0
pydantic-settings==2.0.3
pydantic_core==2.4.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
regex==2023.6.3
requests==2.31.0
requests-file==1.5.1
scipy==1.11.2
sgmllib3k==1.0.0
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
starlette==0.27.0
tinysegmenter==0.3
tldextract==3.4.4
tqdm==4.65.0
typing_extensions==4.7.1
ujson==5.8.0
urllib3==2.0.3
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
